Move data package into conduit.
