+++ conduit/pipeline/pipeline.go
@@ -19,11 +19,11 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/algorand/indexer/conduit"
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
 	"github.com/algorand/indexer/conduit/plugins/importers"
 	"github.com/algorand/indexer/conduit/plugins/processors"
-	"github.com/algorand/indexer/data"
 	"github.com/algorand/indexer/util"
 	"github.com/algorand/indexer/util/metrics"
