+++ conduit/pipeline/pipeline_test.go
@@ -21,11 +21,11 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	"github.com/algorand/indexer/conduit"
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
 	"github.com/algorand/indexer/conduit/plugins/importers"
 	"github.com/algorand/indexer/conduit/plugins/processors"
-	"github.com/algorand/indexer/data"
 	_ "github.com/algorand/indexer/util/metrics"
 )
