+++ conduit/plugins/exporters/example/example_exporter_test.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
-	"github.com/algorand/indexer/data"
 )
 
 var exCons = exporters.ExporterConstructorFunc(func() exporters.Exporter {
