+++ conduit/plugins/exporters/filewriter/file_exporter.go
@@ -12,9 +12,9 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/algorand/indexer/conduit"
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
-	"github.com/algorand/indexer/data"
 	"github.com/algorand/indexer/util"
 )
