+++ conduit/plugins/exporters/noop/noop_exporter_test.go
@@ -7,9 +7,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v3"
 
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
-	"github.com/algorand/indexer/data"
 	testutil "github.com/algorand/indexer/util/test"
 
 	sdk "github.com/algorand/go-algorand-sdk/v2/types"
