+++ conduit/plugins/exporters/postgresql/postgresql_exporter_test.go
@@ -11,10 +11,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v3"
 
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
 	"github.com/algorand/indexer/conduit/plugins/exporters"
 	"github.com/algorand/indexer/conduit/plugins/exporters/postgresql/util"
-	"github.com/algorand/indexer/data"
 	_ "github.com/algorand/indexer/idb/dummy"
 	testutil "github.com/algorand/indexer/util/test"
 )
