+++ conduit/plugins/importers/importer.go
@@ -6,8 +6,8 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/algorand/indexer/conduit"
+	"github.com/algorand/indexer/conduit/data"
 	"github.com/algorand/indexer/conduit/plugins"
-	"github.com/algorand/indexer/data"
 
 	sdk "github.com/algorand/go-algorand-sdk/v2/types"
 )
